Add validator for mute timing and make it provisionable
