+++ pkg/services/ngalert/api/tooling/definitions/alertmanager_validation.go
@@ -2,9 +2,13 @@ package definitions
 
 import (
 	"fmt"
+	"html/template"
+	"regexp"
+	"strings"
 	"time"
 
 	"github.com/prometheus/common/model"
+	"gopkg.in/yaml.v3"
 )
 
 // Validate normalizes a possibly nested Route r, and returns errors if r is invalid.
@@ -52,6 +56,37 @@ func (r *Route) validateChild() error {
 	return nil
 }
 
+func (t *MessageTemplate) Validate() error {
+	if t.Name == "" {
+		return fmt.Errorf("template must have a name")
+	}
+	if t.Template == "" {
+		return fmt.Errorf("template must have content")
+	}
+
+	_, err := template.New("").Parse(t.Template)
+	if err != nil {
+		return fmt.Errorf("invalid template: %w", err)
+	}
+
+	content := strings.TrimSpace(t.Template)
+	found, err := regexp.MatchString(`\{\{\s*define`, content)
+	if err != nil {
+		return fmt.Errorf("failed to match regex: %w", err)
+	}
+	if !found {
+		lines := strings.Split(content, "\n")
+		for i, s := range lines {
+			lines[i] = "  " + s
+		}
+		content = strings.Join(lines, "\n")
+		content = fmt.Sprintf("{{ define \"%s\" }}\n%s\n{{ end }}", t.Name, content)
+	}
+	t.Template = content
+
+	return nil
+}
+
 // Validate normalizes a Route r, and returns errors if r is an invalid root route. Root routes must satisfy a few additional conditions.
 func (r *Route) Validate() error {
 	if len(r.Receiver) == 0 {
@@ -64,4 +99,15 @@ func (r *Route) Validate() error {
 		return fmt.Errorf("root route must not have any mute time intervals")
 	}
 	return r.validateChild()
+}
+
+func (mt *MuteTimeInterval) Validate() error {
+	s, err := yaml.Marshal(mt.MuteTimeInterval)
+	if err != nil {
+		return err
+	}
+	if err = yaml.Unmarshal(s, mt.MuteTimeInterval); err != nil {
+		return err
+	}
+	return nil
 }
