+++ pkg/services/ngalert/api/tooling/definitions/provisioning_templates.go
@@ -1,11 +1,6 @@
 package definitions
 
 import (
-	"fmt"
-	"html/template"
-	"regexp"
-	"strings"
-
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
@@ -76,35 +71,4 @@ func (t *MessageTemplate) ResourceType() string {
 
 func (t *MessageTemplate) ResourceID() string {
 	return t.Name
-}
-
-func (t *MessageTemplate) Validate() error {
-	if t.Name == "" {
-		return fmt.Errorf("template must have a name")
-	}
-	if t.Template == "" {
-		return fmt.Errorf("template must have content")
-	}
-
-	_, err := template.New("").Parse(t.Template)
-	if err != nil {
-		return fmt.Errorf("invalid template: %w", err)
-	}
-
-	content := strings.TrimSpace(t.Template)
-	found, err := regexp.MatchString(`\{\{\s*define`, content)
-	if err != nil {
-		return fmt.Errorf("failed to match regex: %w", err)
-	}
-	if !found {
-		lines := strings.Split(content, "\n")
-		for i, s := range lines {
-			lines[i] = "  " + s
-		}
-		content = strings.Join(lines, "\n")
-		content = fmt.Sprintf("{{ define \"%s\" }}\n%s\n{{ end }}", t.Name, content)
-	}
-	t.Template = content
-
-	return nil
 }
